BUG: TubeTK_TEST_ENVIRONMENT not set
On dashboards, TubeTK_TEST_ENVIRONMENT should be set to OFF
as it is expected that the environment variables on the system
may not have the same value as what the TubeTK configuration
system would expect. This is due to the fact that multiple
versions of TubeTK are typically configured and compiled
on one dashboard machine. This test is only useful for
developers to warn them that something could be wrong on
their machines if they have environment variables set
for ITK_BUILD_DIR and TUBETK_BUILD_DIR.