ENH: Run TestEnvironmentVariables on MicroTumorLinux1_Amazon
The test "TestEnvironmentVariables" requires the environment variables
ITK_BUILD_DIR and TubeTK_BUILD_DIR to be set. These variables are
set in the scripts running ctest because:
1) setting the environment variable on the system (~/.bashrc) require one
more step to configure a new dashboard machine.
2) Cron does not take into account the user environment
3) Setting the variables directly on the system would make the test
pass for one configuration of TubeTK on the computer but not for
the others (Debug-Superbuild, Release-Superbuild, and Slicer).