ENH: Adding TubeTK_TEST_ENVIRONMENT test on MicroTumorWindows1_Amazon
TubeTK_TEST_ENVIRONMENT test fails if environment variables ITK_BUILD_DIR
and TubeTK_BUILD_DIR are not set to value that match the current TubeTK
build directory configuration. This test was disabled on MicroTumorWindows1_Amazon
since the environment variables were not set. The environment variables are
now set in the *bat script that starts the configuration of each build
of the project. TubeTK_TEST_ENVIRONMENT should pass and therefore the test
is enabled.